+++ cluster.tf
@@ -199,6 +199,7 @@ resource "google_container_cluster" "primary" {
     }
   }
 
+
   dynamic "enterprise_config" {
     for_each = var.enterprise_config != null ? [1] : []
     content {
@@ -316,6 +317,14 @@ resource "google_container_cluster" "primary" {
       }
     }
 
+    dynamic "parallelstore_csi_driver_config" {
+      for_each = local.parallelstore_csi_driver_config
+
+      content {
+        enabled = parallelstore_csi_driver_config.value.enabled
+      }
+    }
+
   }
 
   datapath_provider = var.datapath_provider
@@ -412,6 +421,7 @@ resource "google_container_cluster" "primary" {
       machine_type                = lookup(var.node_pools[0], "machine_type", "e2-medium")
       min_cpu_platform            = lookup(var.node_pools[0], "min_cpu_platform", "")
       enable_confidential_storage = lookup(var.node_pools[0], "enable_confidential_storage", false)
+      disk_type                   = lookup(var.node_pools[0], "disk_type", null)
       dynamic "gcfs_config" {
         for_each = lookup(var.node_pools[0], "enable_gcfs", null) != null ? [var.node_pools[0].enable_gcfs] : []
         content {
@@ -489,6 +499,8 @@ resource "google_container_cluster" "primary" {
         enable_secure_boot          = lookup(var.node_pools[0], "enable_secure_boot", false)
         enable_integrity_monitoring = lookup(var.node_pools[0], "enable_integrity_monitoring", true)
       }
+
+      local_ssd_encryption_mode = lookup(var.node_pools[0], "local_ssd_encryption_mode", null)
     }
   }
 
@@ -508,6 +520,14 @@ resource "google_container_cluster" "primary" {
     }
   }
 
+  dynamic "control_plane_endpoints_config" {
+    for_each = var.dns_allow_external_traffic != null ? [1] : []
+    content {
+      dns_endpoint_config {
+        allow_external_traffic = var.dns_allow_external_traffic
+      }
+    }
+  }
 
   remove_default_node_pool = var.remove_default_node_pool
 
@@ -617,10 +637,17 @@ resource "google_container_node_pool" "pools" {
   }
 
   dynamic "network_config" {
-    for_each = length(lookup(each.value, "pod_range", "")) > 0 || lookup(each.value, "enable_private_nodes", null) != null ? [each.value] : []
+    for_each = length(lookup(each.value, "pod_range", "")) > 0 || lookup(each.value, "enable_private_nodes", null) != null || lookup(each.value, "total_egress_bandwidth_tier", null) != null ? [each.value] : []
     content {
       pod_range            = lookup(network_config.value, "pod_range", null)
       enable_private_nodes = lookup(network_config.value, "enable_private_nodes", null)
+
+      dynamic "network_performance_config" {
+        for_each = lookup(network_config.value, "total_egress_bandwidth_tier", "") != "" ? [1] : []
+        content {
+          total_egress_bandwidth_tier = lookup(network_config.value, "total_egress_bandwidth_tier", null)
+        }
+      }
     }
   }
 
@@ -870,6 +897,7 @@ resource "google_container_node_pool" "pools" {
       }
     }
 
+    local_ssd_encryption_mode = lookup(each.value, "local_ssd_encryption_mode", null)
   }
 
   lifecycle {
@@ -933,10 +961,17 @@ resource "google_container_node_pool" "windows_pools" {
   }
 
   dynamic "network_config" {
-    for_each = length(lookup(each.value, "pod_range", "")) > 0 || lookup(each.value, "enable_private_nodes", null) != null ? [each.value] : []
+    for_each = length(lookup(each.value, "pod_range", "")) > 0 || lookup(each.value, "enable_private_nodes", null) != null || lookup(each.value, "total_egress_bandwidth_tier", null) != null ? [each.value] : []
     content {
       pod_range            = lookup(network_config.value, "pod_range", null)
       enable_private_nodes = lookup(network_config.value, "enable_private_nodes", null)
+
+      dynamic "network_performance_config" {
+        for_each = lookup(network_config.value, "total_egress_bandwidth_tier", "") != "" ? [1] : []
+        content {
+          total_egress_bandwidth_tier = lookup(network_config.value, "total_egress_bandwidth_tier", null)
+        }
+      }
     }
   }
 
@@ -1170,6 +1205,7 @@ resource "google_container_node_pool" "windows_pools" {
       }
     }
 
+    local_ssd_encryption_mode = lookup(each.value, "local_ssd_encryption_mode", null)
   }
 
   lifecycle {